[2020 - Day 14] Solved part 1

+++ src/main/kotlin/be/edu/adventofcode/BinaryNumber.kt
@@ -2,10 +2,10 @@ package be.edu.adventofcode
 
 import kotlin.math.pow
 
-fun toDecimal(binaryNumber : String) : Int {
-    var sum = 0
-    binaryNumber.reversed().forEachIndexed {
-            k, v -> sum += v.toString().toInt() * pow(2, k)
+fun toDecimal(binaryNumber: String): Long {
+    var sum = 0L
+    binaryNumber.reversed().forEachIndexed { k, v ->
+        sum += v.toString().toLong() * pow(2, k)
     }
     return sum
 }
@@ -18,4 +18,4 @@ fun fromDecimal(decimalNumber: Int, binaryString: String = "") : String {
     return binaryString.reversed()
 }
 
-fun pow(base: Int, exponent: Int) = base.toDouble().pow(exponent.toDouble()).toInt()
+fun pow(base: Int, exponent: Int) = base.toDouble().pow(exponent.toDouble()).toLong()
